requirements.txt
pyyaml
numpy
pandas
torch
torchaudio
librosa
transformers
scipy
scikit-learn
tqdm
google-cloud-bigquery
pyarrow
matplotlib
seaborn
